requirements.txt
Flask>=2.0.1
Werkzeug>=2.0.1
google-generativeai>=0.3.0
requests>=2.31.0
gunicorn>=20.1.0
pytz
beautifulsoup4
markdown2